Ads Manager v2
Updated some classes on the table from traversy media - not able to figure out whats wrong with my fontawesome icon.  Also it's hard to make it responsive with all those buttons and stuff floating around - still did the basic media query on traversy which doesnt seem to help a lot :( . Also still have to figure out some stuff related to the white border on the search box when in focus and positioning of the buttons. Also I've got no idea how such a complex layout with all these components can be responsive. I think I have to tinker with the positioning as well as the margin of every single components. Or is there a way to simplify this using grid or something . I mean can we use grid for buttons and stuff like that ?